`EuiTableHeaderCellCheckbox` convert to TS
